pyproject.toml
[project]
name = "local-llm-platform"
version = "1.0.0"
description = "Enterprise Local LLM Platform"
authors = [{name = "Your Name", email = "your@email.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    "psycopg2-binary==2.9.9",
    "redis==5.0.1",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "prometheus-fastapi-instrumentator==6.1.0",
    "celery==5.3.4",
    "aioredis==2.0.1",
    "httpx==0.25.1",
    "aiofiles==23.2.1",
    "tenacity==8.2.3",
    "pandas==2.1.3",
    "numpy==1.24.3",
    "scikit-learn==1.3.2",
    "transformers==4.36.0",
    "torch==2.1.0",
    "datasets==2.15.0",
    "peft==0.6.2",
    "accelerate==0.25.0",
    "mlflow==2.9.2",
    "loguru==0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "black==23.11.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "pre-commit==3.5.0",
]
monitoring = [
    "prometheus-client==0.19.0",
    "grafana-api==1.0.4",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
